testing compilation for object restructure
